Migrate system to GHA
