+++ .github/workflows/pr-automation.yml
@@ -1,0 +1,24 @@
+env:
+  GITHUB_TOKEN: ${{ secrets.PULUMI_BOT_TOKEN }}
+jobs:
+  open-pull-request:
+    name: open pull request for ci changes
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout Repo
+      uses: actions/checkout@v2
+    - name: Create Pull Request
+      uses: repo-sync/pull-request@v2
+      with:
+        github_token: ${{ secrets.PULUMI_BOT_TOKEN }}
+        pr_body: "\U0001F680 This PR has been opened because changes have been pushed
+          to ${{ github.ref }}, please review them carefully!"
+        pr_label: automation/pull-request,impact/no-changelog-required
+        pr_reviewer: jaxxstorm,stack72
+        pr_title: "\U0001F916 automated pull-request from pulumi"
+name: pr-automation
+"on":
+  push:
+    branches:
+    - pulumi-automation
+    - automation/pulumi-provider-ci
