+++ .gitignore
@@ -1,5 +1,22 @@
-*.swp
-/vendor/
-/.idea/
+.idea
+.code
+**/vendor/
+.pulumi
+**/bin/
+**/obj/
+Pulumi.*.yaml
+**/node_modules/
+
+**/command-output/
+
+.idea/
 *.iml
-.terraform
+
+yarn.lock
+**/pulumiManifest.go
+
+ci-scripts
+**/schema.go
+**/version.txt
+**/nuget
+**/dist
