+++ .golangci.yml
@@ -1,0 +1,24 @@
+linters:
+  enable:
+  - deadcode
+  - errcheck
+  - goconst
+  - gofmt
+  - golint
+  - gosec
+  - govet
+  - ineffassign
+  - interfacer
+  - lll
+  - megacheck
+  - misspell
+  - nakedret
+  - structcheck
+  - unconvert
+  - varcheck
+  enable-all: false
+run:
+  skip-files:
+  - schema.go
+  - pulumiManifest.go
+  timeout: 10m
