--- .travis.yml
@@ -1,31 +1,0 @@
-# It may be tempting to add parens around each individual clause in this expression, but Travis then builds pushes anyway
-if: branch = master OR branch =~ ^release/ OR branch =~ ^features/ OR tag =~ ^v\d+.*
-language: go
-go: 1.14.x
-git:
-    depth: false
-sudo: true # give us 7.5GB and >2 bursted cores.
-before_install:
-    - git clone https://github.com/pulumi/scripts ${GOPATH}/src/github.com/pulumi/scripts
-    - source ${GOPATH}/src/github.com/pulumi/scripts/ci/prepare-environment.sh
-    - source ${PULUMI_SCRIPTS}/ci/keep-failed-tests.sh
-install:
-    - source ${PULUMI_SCRIPTS}/ci/install-common-toolchain.sh
-    - source ${PULUMI_SCRIPTS}/ci/configure-aws.sh
-    # Use the pulumi-ci AWS profile configured by the above script.
-    # This repo runs tests that require access to create resources in AWS.
-    - export AWS_PROFILE=pulumi-ci
-    - export AWS_SDK_LOAD_CONFIG=1
-    # Install Pulumi
-    - curl -L https://get.pulumi.com/ | bash
-    - export PATH=$HOME/.pulumi/bin:$PATH
-before_script:
-    - ${PULUMI_SCRIPTS}/ci/ensure-dependencies
-after_failure:
-    - ${PULUMI_SCRIPTS}/ci/upload-failed-tests
-notifications:
-    webhooks: https://zlmgkhmhjc.execute-api.us-west-2.amazonaws.com/stage/travis
-jobs:
-    include:
-        - stage: build
-          script: make travis_${TRAVIS_EVENT_TYPE}
