+++ Makefile
@@ -1,88 +1,77 @@
-PROJECT_NAME := Pulumi Terraform Resource Provider
-include build/common.mk
-
 PACK             := terraform
-PACKDIR          := sdk
-NODE_MODULE_NAME := @pulumi/terraform
-NUGET_PKG_NAME   := Pulumi.Terraform
-PROJECT          := github.com/pulumi/pulumi-terraform
-TESTPARALLELISM  := 4
+ORG              := pulumi
+PROJECT          := github.com/${ORG}/pulumi-${PACK}
+NODE_MODULE_NAME := @pulumi/${PACK}
+TF_NAME          := ${PACK}
+PROVIDER_PATH    := provider/v3
+VERSION_PATH     := ${PROVIDER_PATH}/pkg/version.Version
 
-VERSION          ?= $(shell scripts/get-version)
-PYPI_VERSION     := $(shell cd scripts && ./get-py-version)
+PROVIDER        := pulumi-resource-${PACK}
+VERSION         := $(shell pulumictl get version)
 
-VERSION_FLAGS    := -ldflags "-X github.com/pulumi/pulumi-terraform/provider/v3/pkg/version.Version=${VERSION}"
+WORKING_DIR     := $(shell pwd)
 
-DOTNET_PREFIX  := $(firstword $(subst -, ,${VERSION:v%=%})) # e.g. 1.5.0
-DOTNET_SUFFIX  := $(word 2,$(subst -, ,${VERSION:v%=%}))    # e.g. alpha.1
+.PHONY: development provider build_sdks build_nodejs build_dotnet build_go build_python cleanup
 
-ifeq ($(strip ${DOTNET_SUFFIX}),)
-	DOTNET_VERSION := $(strip ${DOTNET_PREFIX})
-else
-	DOTNET_VERSION := $(strip ${DOTNET_PREFIX})-$(strip ${DOTNET_SUFFIX})
-endif
+development:: provider build_sdks install_sdks cleanup # Build the provider & SDKs for a development environment
 
-build::
-	cd provider && go install $(VERSION_FLAGS) ${PROJECT}/provider/v3/cmd/pulumi-resource-${PACK}
-	cd ${PACKDIR}/nodejs/ && \
-		yarn install && \
-		yarn run tsc
-	cp LICENSE ${PACKDIR}/nodejs/package.json ${PACKDIR}/nodejs/yarn.lock \
-		${PACKDIR}/nodejs/bin
-	cp README.md ${PACKDIR}/nodejs/bin/README.md
-	cp README.md ${PACKDIR}/python/README.md
-	sed -i.bak 's/$${VERSION}/$(VERSION)/g' ${PACKDIR}/nodejs/bin/package.json
-	cd ${PACKDIR}/python/ && \
-		$(PYTHON) setup.py clean --all 2>/dev/null && \
-		rm -rf ./bin/ ../python.bin/ && cp -R . ../python.bin && mv ../python.bin ./bin && \
-		sed -i.bak -e "s/\$${VERSION}/$(PYPI_VERSION)/g" -e "s/\$${PLUGIN_VERSION}/$(VERSION)/g" ./bin/setup.py && \
-		cd ./bin && $(PYTHON) setup.py build sdist
-	cd ${PACKDIR}/dotnet/ && \
-		echo "${VERSION:v%=%}" >version.txt && \
-		dotnet build /p:Version=${DOTNET_VERSION}
+# Required for the codegen action that runs in pulumi/pulumi and pulumi/pulumi-terraform-bridge
+build:: provider build_sdks install_sdks
+only_build:: build
 
-lint::
-	#cd provider/cmd/pulumi-resource-terraform && golangci-lint run
+provider:: # build the provider binary
+	(cd provider && go build -a -o $(WORKING_DIR)/bin/${PROVIDER} -ldflags "-X ${PROJECT}/${VERSION_PATH}=${VERSION}" ${PROJECT}/${PROVIDER_PATH}/cmd/${PROVIDER})
 
-install:: build
-	[ ! -e "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)" ] || rm -rf "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)"
-	mkdir -p "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)"
-	cp -r sdk/nodejs/bin/. "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)"
-	rm -rf "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)/node_modules"
-	rm -rf "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)/tests"
-	cd "$(PULUMI_NODE_MODULES)/$(NODE_MODULE_NAME)" && \
-		yarn install --offline --production && \
-		(yarn unlink > /dev/null 2>&1 || true) && \
-		yarn link
-	[ ! -e "$(PULUMI_NUGET)" ] || rm -rf "$(PULUMI_NUGET)/*"
-	find . -name '$(NUGET_PKG_NAME).*.nupkg' -exec cp -p {} ${PULUMI_NUGET} \;
+build_sdks:: provider build_nodejs build_python build_go build_dotnet # build all the sdks
 
-test_fast::
-	cd examples && $(GO_TEST_FAST) .
+build_nodejs:: VERSION := $(shell pulumictl get version --language javascript)
+build_nodejs:: # build the node sdk
+	cd sdk/nodejs/ && \
+        yarn install && \
+        yarn run tsc && \
+        cp ../../README.md ../../LICENSE package.json yarn.lock ./bin/ && \
+    	sed -i.bak -e "s/\$${VERSION}/$(VERSION)/g" ./bin/package.json
 
-test_all::
-	cd examples && $(GO_TEST) .
+build_python:: PYPI_VERSION := $(shell pulumictl get version --language python)
+build_python:: # build the python sdk
+	cd sdk/python/ && \
+        cp ../../README.md . && \
+        python3 setup.py clean --all 2>/dev/null && \
+        rm -rf ./bin/ ../python.bin/ && cp -R . ../python.bin && mv ../python.bin ./bin && \
+        sed -i.bak -e "s/\$${VERSION}/$(PYPI_VERSION)/g" -e "s/\$${PLUGIN_VERSION}/$(VERSION)/g" ./bin/setup.py && \
+        rm ./bin/setup.py.bak && \
+        cd ./bin && python3 setup.py build sdist
 
-.PHONY: publish_tgz
-publish_tgz:
-	$(call STEP_MESSAGE)
-	./scripts/publish_tgz.sh
+build_go:: # build the go sdk
 
-.PHONY: check_clean_worktree
-check_clean_worktree:
-	$$(go env GOPATH)/src/github.com/pulumi/scripts/ci/check-worktree-is-clean.sh
+build_dotnet:: DOTNET_VERSION := $(shell pulumictl get version --language dotnet)
+build_dotnet:: # build the dotnet sdk
+	cd sdk/dotnet/ && \
+		echo "${DOTNET_VERSION}" >version.txt && \
+        dotnet build /p:Version=${DOTNET_VERSION}
 
-# While pulumi-terraform is not built using tfgen/tfbridge, the layout of the source tree is the same as these
-# style of repositories, so we can re-use the common publishing scripts.
-.PHONY: publish_packages
-publish_packages:
-	$(call STEP_MESSAGE)
-	$$(go env GOPATH)/src/github.com/pulumi/scripts/ci/publish-tfgen-package .
-	$$(go env GOPATH)/src/github.com/pulumi/scripts/ci/build-package-docs.sh terraform
+lint_provider:: provider # lint the provider code
+	cd provider && golangci-lint run -c ../.golangci.yml
 
-# The travis_* targets are entrypoints for CI.
-.PHONY: travis_cron travis_push travis_pull_request travis_api
-travis_cron: all
-travis_push: all check_clean_worktree only_test publish_tgz publish_packages
-travis_pull_request: all
-travis_api: all
+cleanup:: # cleans up the temporary directory
+	rm -r $(WORKING_DIR)/bin
+
+help::
+	@grep '^[^.#]\+:\s\+.*#' Makefile | \
+ 	sed "s/\(.\+\):\s*\(.*\) #\s*\(.*\)/`printf "\033[93m"`\1`printf "\033[0m"`	\3 [\2]/" | \
+ 	expand -t20
+
+clean::
+
+install_dotnet_sdk::
+	mkdir -p $(WORKING_DIR)/nuget
+	find . -name '*.nupkg' -print -exec cp -p {} ${WORKING_DIR}/nuget \;
+
+install_python_sdk::
+
+install_go_sdk::
+
+install_nodejs_sdk::
+	yarn link --cwd $(WORKING_DIR)/sdk/nodejs/bin
+
+install_sdks:: install_dotnet_sdk install_python_sdk install_nodejs_sdk
