+++ examples/examples_py_test.go
@@ -1,0 +1,91 @@
+// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+// +build python all
+
+package examples
+
+import (
+	"path"
+	"path/filepath"
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func TestPyLocal013(t *testing.T) {
+	test := getPyBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "localstate-python"),
+			Config: map[string]string{
+				"statefile": "terraform.0-13-0.tfstate",
+			},
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
+func TestPyLocal012(t *testing.T) {
+	test := getPyBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "localstate-python"),
+			Config: map[string]string{
+				"statefile": "terraform.0-12-24.tfstate",
+			},
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
+func TestPyS3013(t *testing.T) {
+	test := getPyBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "s3state-python"),
+			Config: map[string]string{
+				"bucketName": "pulumi-terraform-remote-state-testing",
+				"key":        "0-13-state",
+				"region":     "us-west-2",
+			},
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
+func TestPyS3012(t *testing.T) {
+	test := getPyBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "s3state-python"),
+			Config: map[string]string{
+				"bucketName": "pulumi-terraform-remote-state-testing",
+				"key":        "0-12-state",
+				"region":     "us-west-2",
+			},
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
+func TestPyRemoteBackend(t *testing.T) {
+	test := getPyBaseOptions(t).
+		With(integration.ProgramTestOptions{
+			Dir: path.Join(getCwd(t), "remote-backend-python"),
+			Config: map[string]string{
+				"organization":  getRemoteBackendOrganization(t),
+				"workspaceName": "dev",
+			},
+			Secrets: map[string]string{
+				"tfeToken": getRemoteBackendToken(t),
+			},
+		})
+
+	integration.ProgramTest(t, &test)
+}
+
+func getPyBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions()
+	basePy := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			filepath.Join("..", "sdk", "python", "bin"),
+		},
+	})
+
+	return basePy
+}
