+++ examples/s3state-go/go.mod
@@ -1,7 +1,7 @@
 module s3state-go
 
 require (
-	github.com/pulumi/pulumi-terraform/sdk/v2 v2.0.0-20200514134513-f96b9fce14ef
+	github.com/pulumi/pulumi-terraform/sdk/v3 v3.0.0-20200817232100-bfe03eda7e8b
 	github.com/pulumi/pulumi/sdk/v2 v2.2.1
 )
