--- scripts/go.mod
@@ -1,5 +1,0 @@
-module github.com/pulumi/pulumi-terraform/scripts/v3
-
-go 1.14
-
-require github.com/pulumi/pulumi/sdk/v2 v2.1.0
