+++ sdk/go.mod
@@ -4,6 +4,5 @@ go 1.14
 
 require (
 	github.com/pkg/errors v0.9.1
-	github.com/pulumi/pulumi-terraform/sdk/v2 v2.5.0
 	github.com/pulumi/pulumi/sdk/v2 v2.2.1
 )
